+++ src/inspect_ai/tool/_tools/_web_search/__init__.py
@@ -1,0 +1,3 @@
+from ._web_search import web_search
+
+__all__ = ["web_search"]
