+++ src/inspect_ai/tool/_tools/_web_search/_types.py
@@ -1,0 +1,5 @@
+from typing import Protocol
+
+
+class SearchProvider(Protocol):
+    async def __call__(self, query: str) -> str | None: ...
